structural changes - e.g. list of ingredients
