+++ src/main.py
@@ -1,0 +1,9 @@
+#imports
+from Recipe import Recipe
+
+
+if __name__ == "__main__":
+    a_recipe = Recipe()
+
+
+    print("") #placeholder for later kthxbai
