--- src/recipe.py
@@ -1,24 +1,0 @@
-class Recipe:
-
-    def __init__(self):
-        self.quantity = 0
-        self.unit = ""
-        self.ingredient = ""
-        self.directions = ""
-        self.notes = ""
-        self.name = ""
-        self.synopsis = ""
-
-    def add_recipe(self):
-        return
-
-    def edit_recipe(self):
-        return
-
-    def delete_recipe(self):
-        return
-
-    def display_recipe(self):
-        return
-
-    #we also will want a 'return all recipes' method but this belongs elsewhere
